........
  r22813 | greg | 2007-06-07 07:02:46 -0700 (Thu, 07 Jun 2007) | 2 lines
  
  Return true/false, not o/not-0 for relation checking functions, for those parts of the code that demand "result === false" rather than saying (!result)
........
  r22814 | ivanlanin | 2007-06-07 07:48:26 -0700 (Thu, 07 Jun 2007) | 1 line
  
  Indonesian (id) localisation updates.
........
  r22815 | robchurch | 2007-06-07 08:07:12 -0700 (Thu, 07 Jun 2007) | 3 lines
  
  * Don't require a list of extensions; move it into an option, and default to $wgFileExtensions
  * Let the user know if no suitable files are found
  * Missing release note from r22772
........
  r22816 | amidaniel | 2007-06-07 10:31:08 -0700 (Thu, 07 Jun 2007) | 1 line
  
  (bug 7997) Added ability to Special:Blockip to block users from using Special:Emailuser.
........
  r22817 | tstarling | 2007-06-07 10:53:26 -0700 (Thu, 07 Jun 2007) | 1 line
  
  SqlBagOStuff: Don't write to the database during read-only mode
........
  r22823 | aaron | 2007-06-07 12:52:41 -0700 (Thu, 07 Jun 2007) | 1 line
  
  *Fix comment
........
  r22826 | robchurch | 2007-06-07 13:49:34 -0700 (Thu, 07 Jun 2007) | 2 lines
  
  * (bug 8989) Blacklist 'mhtml' and 'mht' files from upload
  * Trim release note from bug 7997
........
  r22827 | amidaniel | 2007-06-07 14:15:59 -0700 (Thu, 07 Jun 2007) | 1 line
  
  Reworked e-mail blocking. To be enabled for a user, it now requires that $wgSysopEmailBans == true and that the user has the right "blockemail".
........
  r22828 | robchurch | 2007-06-07 15:16:19 -0700 (Thu, 07 Jun 2007) | 3 lines
  
  * (bug 8760) Allow wiki links in "protectexpiry" message
  * Introduce wfMsgWithLinks() and a content-language counterpart, wfMsgForContentWithLinks(); these escape HTML and render wiki links in text, so should be suitable for labels and so on
  * Make Linker::formatLinksInComment() public
........
  r22829 | robchurch | 2007-06-07 17:52:56 -0700 (Thu, 07 Jun 2007) | 1 line
  
  Ditch wfSpecialConfirmemail(), move Special:Confirmemail to the modern special page loading scheme
........
  r22830 | robchurch | 2007-06-07 17:57:11 -0700 (Thu, 07 Jun 2007) | 1 line
  
  Proper credit...
........
  r22831 | greg | 2007-06-07 17:57:22 -0700 (Thu, 07 Jun 2007) | 2 lines
  
  Fix up encodeBlob, don't worry so much about indexInfo compat.
........
  r22832 | greg | 2007-06-07 18:06:52 -0700 (Thu, 07 Jun 2007) | 4 lines
  
  Have Math.php use encodeBlob and decodeBlob when accessing the binary fields math_inputhash and math_outputhash. Better 
  long-term solution is to simply store the hexadecimal as text, but converting via updaters.inc will be complex.
  Fixes bug # 9909.
........
  r22833 | simetrical | 2007-06-07 20:42:55 -0700 (Thu, 07 Jun 2007) | 3 lines
  
  Add 'Brokenlinks' as an alias for 'Wantedpages'.  I suppose this is technically a l10n change, so no RELEASE-NOTES?
  
  Incidentally, when are we going to be implementing the readable special page names (with spaces!) for English?
........
  r22835 | nikerabbit | 2007-06-08 05:57:46 -0700 (Fri, 08 Jun 2007) | 2 lines
  
  * Updates to Slovak (sk) by Helix84
........
  r22837 | nikerabbit | 2007-06-08 06:46:51 -0700 (Fri, 08 Jun 2007) | 2 lines
  
  * Updates to fi
........
  r22839 | robchurch | 2007-06-08 08:16:00 -0700 (Fri, 08 Jun 2007) | 1 line
  
  (bug 5908) Add "DEFAULTSORTKEY" and "DEFAULTCATEGORYSORT" aliases for "DEFAULTSORT" magic word
........
  r22841 | nikerabbit | 2007-06-08 08:27:25 -0700 (Fri, 08 Jun 2007) | 2 lines
  
  * Fixes to fi
........
  r22842 | robchurch | 2007-06-08 08:56:32 -0700 (Fri, 08 Jun 2007) | 2 lines
  
  * (bug 10181) Support the XCache object caching mechanism [patch from Kurt Radwanski]
  * Minor tweak to installer form
........
  r22844 | robchurch | 2007-06-08 09:39:29 -0700 (Fri, 08 Jun 2007) | 1 line
  
  Wrap the recreate-deleted warning dismisser in a CDATA section and run the message through Xml::escapeJsString()
........
  r22845 | robchurch | 2007-06-08 11:12:07 -0700 (Fri, 08 Jun 2007) | 1 line
  
  (bug 9058) Introduce '--aconf' option for all maintenance scripts, to provide a path to the AdminSettings.php file
........
  r22846 | robchurch | 2007-06-08 11:38:36 -0700 (Fri, 08 Jun 2007) | 1 line
  
  (bug 8781) Remind users to check file permissions for LocalSettings.php post-installation
........
  r22847 | ivanlanin | 2007-06-08 11:48:17 -0700 (Fri, 08 Jun 2007) | 1 line
  
  Indonesian (id) localisation updates.
........
  r22851 | robchurch | 2007-06-08 12:52:51 -0700 (Fri, 08 Jun 2007) | 1 line
  
  Whoops, add missing CSS
........
  r22852 | nikerabbit | 2007-06-08 13:25:47 -0700 (Fri, 08 Jun 2007) | 2 lines
  
  * Updates to Romanian (ro) by KlaudiuMihaila, SCriBu, Emily
........


git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/liquidthreads@22857 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ includes/BagOStuff.php
@@ -253,6 +253,9 @@ abstract class SqlBagOStuff extends BagOStuff {
 	}
 
 	function set($key,$value,$exptime=0) {
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$exptime = intval($exptime);
 		if($exptime < 0) $exptime = 0;
 		if($exptime == 0) {
@@ -272,6 +275,9 @@ abstract class SqlBagOStuff extends BagOStuff {
 	}
 
 	function delete($key,$time=0) {
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$this->_query(
 			"DELETE FROM $0 WHERE keyname='$1'", $key );
 		return true; /* ? */
@@ -339,12 +345,18 @@ abstract class SqlBagOStuff extends BagOStuff {
 
 	function expireall() {
 		/* Remove any items that have expired */
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$now = $this->_fromunixtime( time() );
 		$this->_query( "DELETE FROM $0 WHERE exptime < '$now'" );
 	}
 
 	function deleteall() {
 		/* Clear *all* items from cache table */
+		if ( wfReadOnly() ) {
+			return false;
+		}
 		$this->_query( "DELETE FROM $0" );
 	}
 
@@ -550,6 +562,52 @@ class eAccelBagOStuff extends BagOStuff {
 		eaccelerator_unlock( $key );
 		return true;
 	}
+}
+
+/**
+ * Wrapper for XCache object caching functions; identical interface
+ * to the APC wrapper
+ */
+class XCacheBagOStuff extends APCBagOStuff {
+
+	/**
+	 * Get a value from the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @return mixed
+	 */
+	public function get( $key ) {
+		$val = xcache_get( $key );
+		if( is_string( $val ) )
+			$val = unserialize( $val );
+		return $val;
+	}
+	
+	/**
+	 * Store a value in the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @param mixed $value Object to store
+	 * @param int $expire Expiration time
+	 * @return bool
+	 */
+	public function set( $key, $value, $expire = 0 ) {
+		xcache_set( $key, serialize( $value ), $expire );
+		return true;
+	}
+	
+	/**
+	 * Remove a value from the XCache object cache
+	 *
+	 * @param string $key Cache key
+	 * @param int $time Not used in this implementation
+	 * @return bool
+	 */
+	public function delete( $key, $time = 0 ) {
+		xcache_unset( $key );
+		return true;
+	}
+	
 }
 
 /**
